remove game from cart

+++ src/main/java/gov/iti/fusion/services/UserService.java
@@ -14,6 +14,7 @@ public class UserService {
     private final LibraryItemRepository libraryItemRepository;
     private final CartItemRepository cartItemRepository;
     private final WishItemRepository wishItemRepository;
+
     public UserService(HttpServletRequest request) {
         this.userRepository = new UserRepository(request);
         this.libraryItemRepository = new LibraryItemRepository(request);
@@ -21,83 +22,91 @@ public class UserService {
         this.wishItemRepository = new WishItemRepository(request);
 
     }
-    public User save(User user){
+
+    public User save(User user) {
         return userRepository.save(user);
     }
 
-    public User findById(String id){
+    public User findById(String id) {
         return userRepository.findById(User.class, id);
     }
-    public List<User> findAll(){
+
+    public List<User> findAll() {
         return userRepository.findAll(User.class);
     }
-    public User deleteById(String id){
+
+    public User deleteById(String id) {
         return userRepository.deleteById(User.class, id);
     }
 
-    public User findUserbyUsername(String username){
+    public User findUserbyUsername(String username) {
         return userRepository.findUserByUsername(username);
     }
-    public User findUserbyEmail(String email){
+
+    public User findUserbyEmail(String email) {
         return userRepository.findUserByEmail(email);
     }
-    public User findUserbyPhoneNumber(String phoneNumber){
+
+    public User findUserbyPhoneNumber(String phoneNumber) {
         return userRepository.findUserByPhoneNumber(phoneNumber);
     }
 
-    public void addGameToUserLibrary(User user, Game game){
+    public void addGameToUserLibrary(User user, Game game) {
         LibraryItem item = new LibraryItem(user, game);
         libraryItemRepository.save(item);
     }
 
-    public void addGamesToUserLibrary(User user, Collection<Game> games){
-        for(Game game : games){
+    public void addGamesToUserLibrary(User user, Collection<Game> games) {
+        for (Game game : games) {
             addGameToUserLibrary(user, game);
         }
     }
 
-    public void addGameToCart(User user, Game game){
+    public void addGameToCart(User user, Game game) {
         CartItem item = new CartItem(user, game);
         cartItemRepository.save(item);
     }
 
-    public void addGamesToCart(User user, Collection<Game>  games){
-        for(Game game : games){
+    public void addGamesToCart(User user, Collection<Game> games) {
+        for (Game game : games) {
             addGameToCart(user, game);
         }
     }
 
-    public void addGameToWishList(User user, Game game){
+    public void addGameToWishList(User user, Game game) {
         WishItem item = new WishItem(user, game);
         wishItemRepository.save(item);
     }
-    public void deleteGameFromWishList(User user, Game game){
-        WishItem wishItem = wishItemRepository.findWishItem(user.getId(),game.getId());
-        wishItemRepository.deleteById(WishItem.class, wishItem.getId());
-    }
 
-    public void addGamesToWishList(User user, Collection<Game>  games){
-        for(Game game : games){
+    public void addGamesToWishList(User user, Collection<Game> games) {
+        for (Game game : games) {
             addGameToWishList(user, game);
         }
     }
 
-    // public void removeGameFromCart(User user, Game game){
-    //     CartItem cartItem = new CartItem(user, game);
-    //     cartItemRepository.deleteById(CartItem.class, cartItem.getId());
-    // }
+    public void deleteGameFromWishList(User user, Game game) {
+        WishItem wishItem = wishItemRepository.findWishItem(user.getId(), game.getId());
+        wishItemRepository.deleteById(WishItem.class, wishItem.getId());
+    }
 
-    public void removeGamesFromCart(User user){
-       List<CartItem> cartItems = user.getCartItems();
-       for(CartItem cartItem : cartItems)
+    public void removeGameFromCart(User user, Game game) {
+        List<CartItem> cartItems = user.getCartItems();
+        CartItem cartItem = cartItems.stream().filter(c -> c.getGame().getId().equals(game.getId())).findFirst().orElse(null);
+        ;
+        if (cartItem != null)
             cartItemRepository.deleteById(CartItem.class, cartItem.getId());
     }
-    public void removeGamesFromWishList(User user){
-        List<WishItem> wishItems = user.getWishItems();
-        for(WishItem wishItem : wishItems)
-             wishItemRepository.deleteById(WishItem.class, wishItem.getId());
+
+    public void removeGamesFromCart(User user) {
+        List<CartItem> cartItems = user.getCartItems();
+        for (CartItem cartItem : cartItems)
+            cartItemRepository.deleteById(CartItem.class, cartItem.getId());
     }
 
-   
-    
+    public void removeGamesFromWishList(User user) {
+        List<WishItem> wishItems = user.getWishItems();
+        for (WishItem wishItem : wishItems)
+            wishItemRepository.deleteById(WishItem.class, wishItem.getId());
+    }
+
 }
